requirements.txt
geopandas==0.14.3
numpy==2.1.1
pandas==2.2.2
python-dotenv==1.0.1
rasterio==1.3.11
Requests==2.32.3
scipy==1.14.1
Shapely==2.0.6
